frontend: fix include cycle in stores.
