+++ packages/frontend-web/src/app/stores/__spec__/focus.spec.ts
