+++ packages/frontend-web/src/app/stores/categories.ts
@@ -17,7 +17,7 @@ limitations under the License.
 import { Action, createAction, handleActions } from 'redux-actions';
 import { makeTypedFactory, TypedRecord } from 'typed-immutable-record';
 import { ICategoryModel, ModelId } from '../../models';
-import { IAppStateRecord } from './index';
+import { IAppStateRecord } from './appstate';
 
 const STATE_ROOT = ['global', 'categories'];
 const INDEX = [...STATE_ROOT, 'index'];
