+++ packages/frontend-web/src/app/stores/commentSummaryScores.ts
@@ -18,7 +18,7 @@ import { List, Map } from 'immutable';
 import { Action, createAction, handleActions } from 'redux-actions';
 import { makeTypedFactory, TypedRecord} from 'typed-immutable-record';
 import { listCommentSummaryScoresById } from '../platform/dataService';
-import { IAppStateRecord, IThunkAction } from './index';
+import { IAppStateRecord, IThunkAction } from './appstate';
 
 export interface ICommentSummaryScore {
   tagId: string;
