+++ packages/frontend-web/src/app/stores/comments.ts
@@ -19,7 +19,7 @@ import { Action } from 'redux-actions';
 import { ICommentModel } from '../../models';
 import { listCommentsById } from '../platform/dataService';
 import { ILoadCompletePayload, IQueuedModelStateRecord, makeQueuedModelStore } from '../util';
-import { IAppStateRecord, IThunkAction } from './index';
+import { IAppStateRecord, IThunkAction } from './appstate';
 
 const queueModelStore = makeQueuedModelStore<string, ICommentModel>(
   async (commentIds: List<string>) => {
