+++ packages/frontend-web/src/app/stores/taggingSensitivities.ts
@@ -19,7 +19,7 @@ import { Action, createAction, handleActions } from 'redux-actions';
 import { makeTypedFactory, TypedRecord } from 'typed-immutable-record';
 
 import { ITaggingSensitivityModel } from '../../models';
-import { IAppStateRecord } from './index';
+import { IAppStateRecord } from './appstate';
 
 const STATE_ROOT = ['global', 'taggingSensitivities'];
 const TAGGING_SENSITIVITIES_DATA = [...STATE_ROOT, 'items'];
