+++ packages/frontend-web/src/app/stores/tags.ts
@@ -19,7 +19,7 @@ import { Action, createAction, handleActions } from 'redux-actions';
 import { makeTypedFactory, TypedRecord } from 'typed-immutable-record';
 
 import { ITagModel } from '../../models';
-import { IAppStateRecord } from './index';
+import { IAppStateRecord } from './appstate';
 
 const STATE_ROOT = ['global', 'tags'];
 const TAGS_DATA = [...STATE_ROOT, 'items'];
